requirements.txt
beautifulsoup4==4.14.3
python-dotenv==1.2.1
python-telegram-bot[job-queue]==22.6
requests==2.32.5
soupsieve==2.8.3

